I pulsanti erano invertiti, inizio pulizia
