+++ data/skins/default/skin.xml
@@ -394,9 +394,8 @@
     <part id="icon_white"                       x="64" y="256" w="16" h="16" />
     <part id="icon_transparent"                 x="80" y="256" w="16" h="16" />
     <part id="color_wheel_indicator"            x="48" y="192" w="4" h="4" />
-    <part id="no_symmetry"                      x="144" y="240" w="13" h="13" />
-    <part id="horizontal_symmetry"              x="160" y="240" w="13" h="13" />
-    <part id="vertical_symmetry"                x="176" y="240" w="13" h="13" />
+    <part id="vertical_symmetry"                x="160" y="240" w="13" h="13" />
+    <part id="horizontal_symmetry"              x="176" y="240" w="13" h="13" />
     <part id="icon_arrow_down"                  x="144" y="256" w="7" h="4" />
     <part id="icon_close"                       x="152" y="256" w="7" h="7" />
     <part id="icon_search"                      x="160" y="256" w="8" h="8" />
