+++ src/app/tools/symmetry.h
@@ -27,6 +27,24 @@ namespace app {
       virtual void generateStrokes(const Stroke& stroke, Strokes& strokes, ToolLoop* loop) = 0;
     };
 
+    class HorizontalSymmetry : public Symmetry {
+    public:
+      HorizontalSymmetry(int x) : m_x(x) { }
+      void generateStrokes(const Stroke& mainStroke, Strokes& strokes,
+                           ToolLoop* loop) override;
+    private:
+      int m_x;
+    };
+
+    class VerticalSymmetry : public Symmetry {
+    public:
+      VerticalSymmetry(int y) : m_y(y) { }
+      void generateStrokes(const Stroke& mainStroke, Strokes& strokes,
+                           ToolLoop* loop) override;
+    private:
+      int m_y;
+    };
+
   } // namespace tools
 } // namespace app
